Rename `MultiScript` files to use `snake_case`
Consistent with `visual_script` module.
