+++ core/script/SCsub
@@ -3,4 +3,4 @@ Import("env")
 Import("env_goost")
 
 env_goost.add_source_files(env.modules_sources, "*.cpp")
-env_goost.add_source_files(env.modules_sources, "multiscript/*.cpp")
+env_goost.add_source_files(env.modules_sources, "multi_script/*.cpp")
