+++ goost.py
@@ -146,6 +146,7 @@ classes = {
     "LinkedList": "core",
     "ListNode": "core",
     "MultiScript": "script",
+    "MultiScriptOwner": "script",
     "PolyBoolean2D": "geometry",
     "PolyBooleanParameters2D": "geometry",
     "PolyDecomp2D": "geometry",
@@ -189,6 +190,7 @@ class_dependencies = {
     "GoostGeometry2D" : ["PolyBoolean2D", "PolyDecomp2D", "PolyOffset2D"],
     "LightTexture" : "GradientTexture2D",
     "LinkedList" : "ListNode",
+    "MultiScript" : "MultiScriptOwner",
     "PolyBoolean2D" : ["PolyBooleanParameters2D", "PolyNode2D"],
     "PolyDecomp2D" : "PolyDecompParameters2D",
     "PolyOffset2D" : "PolyOffsetParameters2D",
