+++ tests/project/goost/core/script/test_multi_script.gd
